requirements.txt
Django==4.2
djangorestframework==3.14.0
django-redis==5.2.0
redis==4.5.5
django-cors-headers==4.2.0
channels==4.0.0
djangorestframework-simplejwt==5.2.2
bcrypt==4.0.1
setuptools==75.8.0
PyJWT==2.10.1
wfdb==4.2.0
neurokit2==0.2.10
django-cors-headers==4.2.0